// src/types/data.ts
export type DataProps = {
  section: {
    title: string
    description: string
    imageSrc: string
    altImageText: string
  },
  section2: {
    title: string
    description: string
    imageSrc: string
    altImageText: string
  },
  section3: {
    title: string
    description: string
    imageSrc: string
    altImageText: string
    incrementSpacingBottom: boolean
  }
}